# mimirpb: optimize label sorting in timeseries type