--- a/pkg/mimirpb/timeseries_test.go
+++ b/pkg/mimirpb/timeseries_test.go
@@ -7,6 +7,7 @@ package mimirpb
 
 import (
 	"crypto/rand"
+	"fmt"
 	"reflect"
 	"sort"
 	"testing"
@@ -424,4 +425,43 @@ func TestPreallocTimeseries_SetLabels(t *testing.T) {
 
 	require.Equal(t, []LabelAdapter{{Name: "__name__", Value: "hello"}, {Name: "lbl", Value: "world"}}, p.Labels)
 	require.Nil(t, p.marshalledData)
+}
+
+func BenchmarkPreallocTimeseries_SortLabelsIfNeeded(b *testing.B) {
+	bcs := []int{10, 100, 1_000, 10_000, 100_000, 1_000_000}
+
+	for _, lbCount := range bcs {
+		b.Run(fmt.Sprintf("num_labels=%d", lbCount), func(b *testing.B) {
+			// Generate unordered labels set.
+			lbs := make([]labels.Label, 0, lbCount)
+			for i := 0; i < lbCount; i++ {
+				lbName := fmt.Sprintf("lbl_%d", lbCount-i)
+				lbValue := fmt.Sprintf("val_%d", lbCount-i)
+				lbs = append(lbs, labels.Label{Name: lbName, Value: lbValue})
+			}
+			unorderedLabels := FromLabelsToLabelAdapters(lbs)
+
+			p := PreallocTimeseries{
+				TimeSeries: &TimeSeries{},
+			}
+
+			// Copy unordered labels set for each benchmark iteration.
+			benchmarkUnorderedLabels := make([][]LabelAdapter, b.N)
+			for i := 0; i < b.N; i++ {
+				benchmarkLabels := make([]LabelAdapter, len(unorderedLabels))
+				copy(benchmarkLabels, unorderedLabels)
+				benchmarkUnorderedLabels[i] = benchmarkLabels
+			}
+
+			// Warmup.
+			b.ResetTimer()
+			b.ReportAllocs()
+
+			// Run benchmark.
+			for i := 0; i < b.N; i++ {
+				p.SetLabels(benchmarkUnorderedLabels[i])
+				p.SortLabelsIfNeeded()
+			}
+		})
+	}
 }