@pstibrany 2023-06-23
That method is optimized for case when sorting is not needed -- which is default when data is received from Prometheus or agents. There were even discussions about dropping unsorted labels in the past as invalid. Have you considered fixing the source of unsorted labels?